Created project's solution, updated license and preview
