--- .vscode/tasks.json
@@ -1,17 +1,0 @@
-{
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "build g++",
-            "type": "shell",
-            "command": "g++",
-            "args": [
-                "main.cpp",
-                "mars.cpp",
-                "rover.cpp",
-                "types.cpp"
-            ],
-            "group": "build"
-        }
-    ]
-}
