+++ RoverOnMars/RoverOnMars.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35931.197 d17.13
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RoverOnMars", "RoverOnMars.vcxproj", "{6767ABBA-C0BD-4C85-994F-5510B7AA781B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Debug|x64.ActiveCfg = Debug|x64
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Debug|x64.Build.0 = Debug|x64
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Debug|x86.ActiveCfg = Debug|Win32
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Debug|x86.Build.0 = Debug|Win32
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Release|x64.ActiveCfg = Release|x64
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Release|x64.Build.0 = Release|x64
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Release|x86.ActiveCfg = Release|Win32
+		{6767ABBA-C0BD-4C85-994F-5510B7AA781B}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {11B71987-2DF8-4A33-BA33-C2A56BDCA88A}
+	EndGlobalSection
+EndGlobal
