+++ RoverOnMars/mars.cpp
@@ -10,24 +10,24 @@
 
 void Mars::init()
 {
-    char objects[] = {' ', ' ', ' ', ' ', ' ', ' ', 'X', '#', '@', '$' };
-    int numOfObjects = (sizeof(objects)/sizeof(*objects)); 
+    char objects[] = { ' ', ' ', ' ', ' ', ' ', ' ', 'X', '#', '@', '$' };
+    int numOfObjects = (sizeof(objects) / sizeof(*objects));
 
     // TODO : Size to be random generated, perhaps multiplication of 5
     m_DimX = 15;
     m_DimY = 15;
 
     // Allocation
-    map.resize(m_DimY); 
-    for(int i = 0 ; i < m_DimY ; ++i)
+    map.resize(m_DimY);
+    for (int i = 0; i < m_DimY; ++i)
     {
         map[i].resize(m_DimX); //resize each rows
     }
 
     // Map object initialization
-    for (int i = 0; i <m_DimY; ++i)
+    for (int i = 0; i < m_DimY; ++i)
     {
-        for (int j = 0 ; j < m_DimX; ++j)
+        for (int j = 0; j < m_DimX; ++j)
         {
             int objectID = std::rand() % numOfObjects;
             map[i][j] = objects[objectID];
@@ -37,11 +37,11 @@ void Mars::init()
 
 void Mars::display()
 {
-    if(m_GameOver)
+    if (m_GameOver)
     {
         system("cls");
         std::cout << "Game Over" << std::endl;
-        std::cout<<"Gold Count: " <<m_DisplayScore << std::endl;
+        std::cout << "Gold Count: " << m_DisplayScore << std::endl;
         return;
     }
 
@@ -50,15 +50,15 @@ void Mars::display()
     std::cout << " -------------------------------" << std::endl;
     std::cout << " = Curiosity, Welcome to Mars! =" << std::endl;
     std::cout << " -------------------------------" << std::endl;
-    for (int i = 0 ; i < m_DimY; ++i)
+    for (int i = 0; i < m_DimY; ++i)
     {
         std::cout << "  "; drawSeparator();
 
         // Row Index Handling
         // Fill in a space for digit less than 10 that causes alignment on rows.
         const int rowIndex = m_DimY - i;
-        if(rowIndex < 10) 
-            std::cout<< " ";
+        if (rowIndex < 10)
+            std::cout << " ";
         std::cout << std::setw(1) << rowIndex;
 
         drawRow(i);
@@ -67,14 +67,14 @@ void Mars::display()
     // Reserve space for index then start draw boarder
     std::cout << "  "; drawSeparator();
 
-     // Column Index Handling - Tenth digit
+    // Column Index Handling - Tenth digit
     std::cout << "  ";
     for (int j = 0; j < m_DimX; ++j)
     {
         int digit = (j + 1) / 10;
         std::cout << " ";
 
-        if(digit == 0)
+        if (digit == 0)
         {
             std::cout << " ";
         }
@@ -87,17 +87,17 @@ void Mars::display()
 
     // Column Index Handling - Single digit
     std::cout << "  ";
-    for (int j=0; j< m_DimX; ++j)
+    for (int j = 0; j < m_DimX; ++j)
     {
-        std::cout << " " << (j + 1) %10;
+        std::cout << " " << (j + 1) % 10;
     }
     std::cout << std::endl << std::endl;
 
-    std::cout<<std::endl;
-    std::cout<<"Gold Count: " <<m_DisplayScore << std::endl;
+    std::cout << std::endl;
+    std::cout << "Gold Count: " << m_DisplayScore << std::endl;
 }
 
-char Mars::getObject(const Point& point) const 
+char Mars::getObject(const Point& point) const
 {
     return map[point.y][point.x];
 }
@@ -151,20 +151,20 @@ void Mars::drawRow(const int i)
 {
     for (int j = 0; j < m_DimX; ++j)
     {
-        char display = '?'; 
-        
+        char display = ' ';
+
         Point p = Point(j, i);
         const bool isPlayer = (m_RoverLocation == p);
-        const bool isFrontRow = matchPoints(p, EnumUtil::getPointRowAheadOf(m_RoverLocation, m_RoverFacing)); 
+        const bool isFrontRow = matchPoints(p, EnumUtil::getPointRowAheadOf(m_RoverLocation, m_RoverFacing));
         if (isPlayer)
         {
             display = EnumUtil::toSymbol(m_RoverFacing);
         }
-        else if(isExplored(p) || m_IsDebug) 
+        else if (isExplored(p) || m_IsDebug)
         {
             display = map[i][j];
         }
-        
+
         std::cout << "|" << display;
     }
     std::cout << "|" << std::endl;
@@ -172,9 +172,9 @@ void Mars::drawRow(const int i)
 
 bool Mars::matchPoints(const Point& point, std::vector<Point> points)
 {
-    for(Point& p : points)
+    for (Point& p : points)
     {
-        if(p == point)
+        if (p == point)
         {
             return true;
         }
@@ -185,9 +185,9 @@ bool Mars::matchPoints(const Point& point, std::vector<Point> points)
 
 bool Mars::isExplored(const Point& point)
 {
-    for(auto&& p : *m_MapReference)
+    for (auto&& p : *m_MapReference)
     {
-        if(p == point)
+        if (p == point)
         {
             return true;
         }
